Use the core component in healthcare
