+++ packages/design-system/src/components/PrivacySettingsDialog/PrivacySettingsDialog.tsx
@@ -1,7 +1,9 @@
 import React from 'react';
-import { Button, Dialog, TFunction } from '@cmsgov/design-system';
-import { getPrivacySettings, setPrivacySettings } from './privacySettings';
 import PrivacySettingsTable from './PrivacySettingsTable';
+import { Button } from '../Button/index';
+import { Dialog } from '../Dialog/index';
+import { getPrivacySettings, setPrivacySettings } from './privacySettings';
+import { t } from '../i18n';
 import { useState } from 'react';
 
 const privacySettingConfigs = [
@@ -11,8 +13,10 @@ const privacySettingConfigs = [
 ];
 
 interface PrivacySettingsDialogProps {
+  domain: string;
+  privacyPolicyUrl: string;
+  thirdPartyPoliciesUrl?: string;
   onExit: () => void;
-  t: TFunction;
 }
 
 export const PrivacySettingsDialog = (props: PrivacySettingsDialogProps) => {
@@ -30,16 +34,23 @@ export const PrivacySettingsDialog = (props: PrivacySettingsDialogProps) => {
     props.onExit();
   }
 
-  const { t, ...dialogProps } = props;
+  const { domain, privacyPolicyUrl, thirdPartyPoliciesUrl, ...dialogProps } = props;
   const privacySettingsProperties = privacySettingConfigs.map((config) => ({
     ...config,
     value: localPrivacySettings[config.settingsKey],
   }));
 
+  let intro = t('privacy.introText', { domain });
+  intro += ' ' + t('privacy.privacyPolicy', { url: privacyPolicyUrl });
+  if (thirdPartyPoliciesUrl) {
+    intro += ' ' + t('privacy.thirdPartyPolicies', { url: thirdPartyPoliciesUrl });
+  }
+  intro += '.';
+
   return (
     <Dialog
       {...dialogProps}
-      heading={t('privacy.dialogTitle')}
+      heading={t('privacy.dialogTitle', { domain })}
       size="full"
       actions={
         <Button variation="solid" onClick={savePrivacySettings}>
@@ -47,10 +58,10 @@ export const PrivacySettingsDialog = (props: PrivacySettingsDialogProps) => {
         </Button>
       }
     >
-      <p dangerouslySetInnerHTML={{ __html: t('privacy.introText') }} />
+      <p dangerouslySetInnerHTML={{ __html: intro }} />
 
       <PrivacySettingsTable
-        t={t}
+        domain={domain}
         privacySettings={privacySettingsProperties}
         setPrivacySetting={setPrivacySetting}
       />
