+++ packages/design-system/src/components/PrivacySettingsDialog/index.ts
@@ -1,0 +1,1 @@
+export * from './PrivacySettingsDialog';
