+++ packages/design-system/src/components/index.ts
@@ -16,6 +16,7 @@ export * from './InlineError';
 export * from './MonthPicker';
 export * from './NativeDialog';
 export * from './Pagination';
+export * from './PrivacySettingsDialog';
 export * from './Review';
 export * from './SkipNav';
 export * from './Spinner';
