+++ packages/ds-healthcare-gov/src/components/Footer/PrivacySettingsLink.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import PrivacySettingsDialog from '../PrivacySettings/PrivacySettingsDialog';
+import { PrivacySettingsDialog } from '@cmsgov/design-system';
 import { useState } from 'react';
 
 import { tWithLanguage } from '../i18n';
@@ -11,11 +11,8 @@ interface PrivacySettingsLinkProps {
 
 export const PrivacySettingsLink = (props: PrivacySettingsLinkProps) => {
   const [showDialog, setShowDialog] = useState(false);
-
   const t = tWithLanguage();
-
   const openDialog = () => setShowDialog(true);
-
   const closeDialog = () => setShowDialog(false);
 
   return (
@@ -23,7 +20,14 @@ export const PrivacySettingsLink = (props: PrivacySettingsLinkProps) => {
       <button className={props.className} onClick={openDialog}>
         {props.children || t('footer.privacySettings')}
       </button>
-      {showDialog && <PrivacySettingsDialog onExit={closeDialog} t={t} />}
+      {showDialog && (
+        <PrivacySettingsDialog
+          onExit={closeDialog}
+          domain="HealthCare.gov"
+          privacyPolicyUrl="https://www.healthcare.gov/privacy/"
+          thirdPartyPoliciesUrl="https://www.healthcare.gov/third-party-privacy-policies/"
+        />
+      )}
     </>
   );
 };
