+++ packages/ds-healthcare-gov/src/styles/components/index.scss
@@ -4,4 +4,3 @@
 @use 'Header';
 @use 'Inset';
 @use 'Logo';
-@use 'PrivacySettingsTable';
